Init commit

+++ src/optimizer/snopt/SnoptInterface.h
@@ -28,42 +28,42 @@ namespace optimizer {
             enum AbnormalType { NONE = 0, INFEASIBLE = 1, HESSIAN_UPDATE = 2, HESSIAN_RESET = 3 }; 
 
             struct Slack {
-                int constr_idx;
+                long constr_idx;
                 SlackType type;
                 double val;
             };
 
             typedef int (*updateFunc) (long mask, int compute_gradients, double *coef_values, void *update_data);
 		
-            SnoptInterface(int constr_total, int coef_total, 
-                           int nonlin_constr_total, int nonlin_obj_coef, int nonlin_jac_coef,
-                           int *constr_eqns, int has_objective, VVD J, VVB JMap, std::vector<double> *constraints,
+            SnoptInterface(long constr_total, long coef_total, 
+                           long nonlin_constr_total, long nonlin_obj_coef, long nonlin_jac_coef,
+                           long *constr_eqns, long has_objective, VVD J, VVB JMap, std::vector<double> *constraints,
                            double *objective, std::vector<double> *gradient,	updateFunc update_f, void *update_d);
 		
-            SnoptInterface(int has_objective, VVD J, VVB JMap, std::vector<double> *constraints, double *objective,
+            SnoptInterface(long has_objective, VVD J, VVB JMap, std::vector<double> *constraints, double *objective,
                            std::vector<double> *gradient, updateFunc update_f, void *update_d);
             ~SnoptInterface();
 
-            Return solve(double *x, double *lo_bounds, double *hi_bounds, int unit = 4);
+            Return solve(double *x, double *lo_bounds, double *hi_bounds, long unit = 4);
 
-            void clear(long mask, int compute_derivs);
+            void clear(long mask, long compute_derivs);
 
-            void resizeJacobian(int coef_total, int nonlin_coef_total, int constr_total, int nonlin_constr_total);
+            void resizeJacobian(long coef_total, long nonlin_coef_total, long constr_total, long nonlin_constr_total);
             void resizeCoef();
 
-            int mNumConstr;
-            int mNumCoef;
-            int mNumNonlinConstr;
-            int mNumNonlinObjCoef;
-            int mNumNonlinJacCoef;
+            long mNumConstr;
+            long mNumCoef;
+            long mNumNonlinConstr;
+            long mNumNonlinObjCoef;
+            long mNumNonlinJacCoef;
 
             double *mSolverX;
             double *mProblemX;
             double *mBoundsLo;
             double *mBoundsHi;
-            int *mConstrEqns;
+            long *mConstrEqns;
 
-            int mHasObjective;
+            long mHasObjective;
 
             static SnoptInterface *mRef;
 
@@ -79,35 +79,35 @@ namespace optimizer {
             VVB mCoefMap;
             double mReturnedObj;
 
-            int mOutput;
-            int mSum;
+            long mOutput;
+            long mSum;
             bool mCheckTerm;
             bool mTermination;
             AbnormalType mAbnormal;
-            int mBreak;
+            long mBreak;
 
             void updateSolverX();
-            void update(long mask, int compute_derivs, double *x);
-            static void checkTermination(int *iAbort, double *xs);
+            void update(long mask, long compute_derivs, double *x);
+            static void checkTermination(long *iAbort, double *xs);
 
 
         protected:
             SnoptInterface::updateFunc mUpdateFunc;
             void *mUpdateData;
 
-            void scaleValues(long update_type, int compute_derivs);
+            void scaleValues(long update_type, long compute_derivs);
 
         private:
-            static void snoptObj(int *mode, int *nn_obj, double *x, 
-                                 double *f_obj, double *g_obj, int *nstate, 
-                                 char *cu, int *lencu, int *iu, int *leniu, 
-                                 double *ru, int *lenru);
-            static void snoptJac(int *mode, int *nn_con, int *nn_jac, int *ne_jac,
-                                 double *x, double *f_con, double *g_con, int *nstate, 
-                                 char *cu, int *lencu, int *iu, int *leniu, 
-                                 double *ru, int *lenru);
-            void fillUpSnoptFormat(VVD jacobian, double **a, int **ha, int **ka);
-            int sparseCount(int col);
+            static void snoptObj(long *mode, long *nn_obj, double *x, 
+                                 double *f_obj, double *g_obj, long *nstate, 
+                                 char *cu, long *lencu, long *iu, long *leniu, 
+                                 double *ru, long *lenru);
+            static void snoptJac(long *mode, long *nn_con, long *nn_jac, long *ne_jac,
+                                 double *x, double *f_con, double *g_con, long *nstate, 
+                                 char *cu, long *lencu, long *iu, long *leniu, 
+                                 double *ru, long *lenru);
+            void fillUpSnoptFormat(VVD jacobian, double **a, long **ha, long **ka);
+            long sparseCount(long col);
 
         };
